chore: mirror project structure in tests/
Reorganize tests so file paths reflect subject + lifecycle:
  tests/unit/      — no network, runs in CI without secrets
  tests/integration/ — hits live API, skips without credentials

Subject-based layout means a failing `tests/unit/cli/lib/api/auth.test.mjs`
points straight at `cli/lib/api/auth.js`. Grab-bag files (unit.test.mjs,
cli.test.mjs, cli-routing.test.mjs) are split by subject.

Main moves:
- tests/auth.test.mjs             → tests/unit/cli/lib/api/auth.test.mjs
- tests/wallet-config.test.mjs    → tests/unit/cli/lib/config.test.mjs
- tests/wallet-ows.test.mjs       → tests/unit/cli/lib/wallet/keystore.test.mjs
- tests/consistency.test.mjs      → tests/unit/consistency.test.mjs
- tests/examples.test.mjs         → tests/unit/examples.test.mjs
- tests/skills.test.mjs           → tests/unit/skills.test.mjs
- tests/tool-catalog.test.mjs     → tests/unit/mcp-tools.test.mjs
- tests/integration.test.mjs      → tests/integration/api.test.mjs
- tests/auth.integration.test.mjs → tests/integration/auth.test.mjs

Content splits (new files absorb subjects from deleted ones):
- tests/unit.test.mjs split into cli/lib/util/{flags,validate,analyze}.test.mjs
  plus its basicAuthHeader suite merged into cli/lib/api/auth.test.mjs.
- tests/wallet-flags.test.mjs merged into cli/lib/util/flags.test.mjs.
- tests/cli.test.mjs renamed to cli/zerion.test.mjs; its live `/chains/`
  test moves to tests/integration/chains.test.mjs.
- tests/cli-routing.test.mjs renamed to cli/router.test.mjs; its two live
  `/chains/` tests move to tests/integration/chains.test.mjs.

Import cleanup:
- package.json gains `"imports": { "#zerion-ai/*": "./*" }` so tests
  reference source as `#zerion-ai/cli/lib/api/auth.js` instead of
  `../../../../cli/lib/api/auth.js`. Works from any depth.
- npm scripts switched from file enumeration to quoted globs:
  `node --test 'tests/unit/**/*.test.mjs'` (quotes let Node's own
  glob matcher resolve the pattern rather than the shell).

Counts before/after: 200 → 208 tests (net +8 from merging wallet-flags
extra cases and the new integration/chains.test.mjs split).
