+++ package.json
@@ -14,11 +14,15 @@
     "LICENSE"
   ],
   "scripts": {
-    "test": "node --test tests/*.test.mjs",
-    "test:unit": "node --test tests/unit.test.mjs tests/tool-catalog.test.mjs tests/examples.test.mjs tests/consistency.test.mjs tests/auth.test.mjs",
-    "test:integration": "node --test tests/integration.test.mjs tests/auth.integration.test.mjs",
-    "test:auth": "node --test tests/auth.integration.test.mjs",
+    "test": "node --test 'tests/**/*.test.mjs'",
+    "test:unit": "node --test 'tests/unit/**/*.test.mjs'",
+    "test:integration": "node --test 'tests/integration/**/*.test.mjs'",
+    "test:integration:auth": "node --test tests/integration/auth.test.mjs",
     "prepublishOnly": "npm test"
+  },
+  "imports": {
+    "#zerion-ai/*": "./*",
+    "#zerion-ai": "./package.json"
   },
   "engines": {
     "node": ">=20"
