+++ tests/integration/auth.test.mjs
@@ -15,15 +15,35 @@
 // The endpoint under test is /wallets/{addr}/transactions/?page[size]=1
 // (via `zerion history <addr> --limit 1`) — chosen to minimize response
 // size and keep each call cheap for the API side.
+//
+// How to run a single mode. ZERION_API_BASE is optional — set it to point
+// at a port-forwarded or custom environment; omit to use the default
+// (https://api.zerion.io/v1).
+//
+//   # apiKey (free — no on-chain payment)
+//   ZERION_API_KEY=<key> ZERION_API_BASE=http://localhost:8000/v1 node --test --test-name-pattern="apiKey" tests/integration/auth.test.mjs
+//
+//   # x402 on Base
+//   ZERION_TEST_EVM_KEY=0x... ZERION_API_BASE=http://localhost:8000/v1 node --test --test-name-pattern="x402 on Base" tests/integration/auth.test.mjs
+//
+//   # x402 on Solana
+//   ZERION_TEST_SOLANA_KEY=<base58> ZERION_API_BASE=http://localhost:8000/v1 node --test --test-name-pattern="x402 on Solana" tests/integration/auth.test.mjs
+//
+//   # MPP on Tempo
+//   ZERION_TEST_TEMPO_KEY=0x... ZERION_API_BASE=http://localhost:8000/v1 node --test --test-name-pattern="MPP" tests/integration/auth.test.mjs
+//
+//   # all four modes together (~$0.03 total)
+//   ZERION_API_KEY=<key> ZERION_TEST_EVM_KEY=0x... ZERION_TEST_SOLANA_KEY=<base58> ZERION_TEST_TEMPO_KEY=0x... ZERION_API_BASE=http://localhost:8000/v1 npm run test:integration:auth
+//
+// Each mode skips independently if its key is missing, so partial env
+// sets (e.g., just ZERION_API_KEY + ZERION_TEST_EVM_KEY) are fine.
 
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 import { execFile } from "node:child_process";
 import { fileURLToPath } from "node:url";
-import { dirname, join } from "node:path";
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const BIN = join(__dirname, "../cli/zerion.js");
+const BIN = fileURLToPath(import.meta.resolve("#zerion-ai/cli/zerion.js"));
 
 const VITALIK = "0x42b9dF65B219B3dD36FF330A4dD8f327A6Ada990";
 const HISTORY_ARGS = ["history", VITALIK, "--limit", "1"];
