+++ tests/unit/cli/lib/config.test.mjs
@@ -20,7 +20,7 @@ describe("config", () => {
     }
     mkdirSync(testDir, { recursive: true });
     // Fresh import each time
-    config = await import("../cli/lib/config.js");
+    config = await import("#zerion-ai/cli/lib/config.js");
   });
 
   afterEach(() => {
